invest/requirements.txt
pyqlib>=0.9.5
lightgbm>=4.0
scikit-learn>=1.3
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
